+++ CHANGELOG.md
@@ -45,6 +45,9 @@ v2.6.0 [unreleased]
   ```
 - [#537](https://github.com/latera/homs/pull/537) Add fileinput field name to uploaded file metadata.
 
+### Bugfixes
+- [#538](https://github.com/latera/homs/pull/538) Removed usage of rubocop-junit_formatter gem.
+
 v2.5.1 [2021-04-08]
 ### Bugfixes
 - [#526](https://github.com/latera/homs/pull/526) Fix fetch BP translations.